Limit number of identical DNS queries we will reply to.
If the same query arrives from second and subsequent requestors
whilst we are in progress of answering it from upstream, we
queue the extra requests and answer them all once the answer is known.

This fix limits the number of copies of the query (to 20). More than that
bumps earlier queries in favour of later ones. This mitigates DoS scenarios.